pyproject.toml
[tool.poetry]
name = "hugging-chat-app"
version = "0.1.0"
description = ""
authors = ["sucream <ghdskatnghds@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.9.3"
python-dotenv = "^1.0.1"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
